feat: styled logo
a logo that will take us to our home page - more importantly though, it's styled.

the styling includes the main attributes we want for the header, its internal anchor

it also has a media query because when we shrink the window, we want the logo to be centered on the page.