feat: add image upload
integrating with cloudinary, this change makes a post request to
retrieve the urls from Cloudinary's servers so that we can store them
along with the rest of the item list when we create the item.

future enhancements would be to move this to the backend so that we only
store images on cloudinary _once_ the form's submitted.