feat: add pagination styling
before getting into modifying the queries for the items page, we have
added a pagination style which will tab between items pages by using
the page query's props (in this case, the page).

In this case `query` in the props refers to the query parameters in the
URL _not_ from a fetch request.

This is what we initially set up in the _app.js with the getInitialProps
.