feat: user component
this render prop user component can be used to retrieve a user's
identity.

this also takes advantage of some magic -- this request doesn't need to
supply the userid since:
a) it is stored in cookies
b) the backend middleware will sniff for the userid which will be needed
to execute the query `identity.